[cleanup] removing dependency on console project from tests by moving ParseNamespace method to CodeUtilities class
