+++ XmlSchemaClassGenerator.Tests/XmlSchemaClassGenerator.Tests.csproj
@@ -21,7 +21,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\XmlSampleGenerator\XmlSampleGenerator.csproj" />
-    <ProjectReference Include="..\XmlSchemaClassGenerator.Console\XmlSchemaClassGenerator.Console.csproj" />
     <ProjectReference Include="..\XmlSchemaClassGenerator\XmlSchemaClassGenerator.csproj" />
   </ItemGroup>
   <ItemGroup>
