+++ XmlSchemaClassGenerator/CodeUtilities.cs
@@ -297,5 +297,20 @@ namespace XmlSchemaClassGenerator
         };
 
         internal static Uri CreateUri(string uri) => string.IsNullOrEmpty(uri) ? null : new Uri(uri);
+
+        public static KeyValuePair<NamespaceKey, string> ParseNamespace(string nsArg, string namespacePrefix)
+        {
+            var parts = nsArg.Split(new[] { '=' }, 2);
+            var xmlNs = parts[0];
+            var netNs = parts[1];
+            var parts2 = xmlNs.Split(new[] { '|' }, 2);
+            var source = parts2.Length == 2 ? new Uri(parts2[1], UriKind.RelativeOrAbsolute) : null;
+            xmlNs = parts2[0];
+            if (!string.IsNullOrEmpty(namespacePrefix))
+            {
+                netNs = namespacePrefix + "." + netNs;
+            }
+            return new KeyValuePair<NamespaceKey, string>(new NamespaceKey(source, xmlNs), netNs);
+        }
     }
 }
